+++ sdk/resourcemanager/compute/armcompute/options.go
@@ -1174,6 +1174,13 @@ type VirtualMachineImagesClientListSKUsOptions struct {
 	// placeholder for future optional parameters
 }
 
+// VirtualMachineImagesClientListWithPropertiesOptions contains the optional parameters for the VirtualMachineImagesClient.ListWithProperties
+// method.
+type VirtualMachineImagesClientListWithPropertiesOptions struct {
+	Orderby *string
+	Top     *int32
+}
+
 // VirtualMachineImagesEdgeZoneClientGetOptions contains the optional parameters for the VirtualMachineImagesEdgeZoneClient.Get
 // method.
 type VirtualMachineImagesEdgeZoneClientGetOptions struct {